# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CardLister API - trading card eBay listing tool
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * @nullable
 */
export type CardHoloType = typeof CardHoloType[keyof typeof CardHoloType] | null;


export const CardHoloType = {
  standard: 'standard',
  holo: 'holo',
  reverse_holo: 'reverse_holo',
} as const;

export type CardStatus = typeof CardStatus[keyof typeof CardStatus];


export const CardStatus = {
  pending: 'pending',
  reviewed: 'reviewed',
  listed: 'listed',
} as const;

export interface Card {
  id: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  cardName?: string | null;
  /** @nullable */
  setName?: string | null;
  /** @nullable */
  cardNumber?: string | null;
  /** @nullable */
  year?: string | null;
  /**
     * NM, LP, MP, HP, or D
     * @nullable
     */
  quality?: string | null;
  /** @nullable */
  holoType?: CardHoloType;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  rarity?: string | null;
  /** @nullable */
  notes?: string | null;
  status: CardStatus;
  /** @nullable */
  suggestedPrice?: number | null;
  /** @nullable */
  ebayListingId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CardInputHoloType = typeof CardInputHoloType[keyof typeof CardInputHoloType];


export const CardInputHoloType = {
  standard: 'standard',
  holo: 'holo',
  reverse_holo: 'reverse_holo',
} as const;

export interface CardInput {
  imageUrl?: string;
  cardName?: string;
  setName?: string;
  cardNumber?: string;
  year?: string;
  quality?: string;
  holoType?: CardInputHoloType;
  language?: string;
  rarity?: string;
  notes?: string;
  suggestedPrice?: number;
}

export type CardUpdateHoloType = typeof CardUpdateHoloType[keyof typeof CardUpdateHoloType];


export const CardUpdateHoloType = {
  standard: 'standard',
  holo: 'holo',
  reverse_holo: 'reverse_holo',
} as const;

export type CardUpdateStatus = typeof CardUpdateStatus[keyof typeof CardUpdateStatus];


export const CardUpdateStatus = {
  pending: 'pending',
  reviewed: 'reviewed',
  listed: 'listed',
} as const;

export interface CardUpdate {
  imageUrl?: string;
  cardName?: string;
  setName?: string;
  cardNumber?: string;
  year?: string;
  quality?: string;
  holoType?: CardUpdateHoloType;
  language?: string;
  rarity?: string;
  notes?: string;
  status?: CardUpdateStatus;
  suggestedPrice?: number;
}

export interface CardAnalyzeInput {
  /** Base64-encoded front image data */
  imageBase64: string;
  /** Base64-encoded back image data (optional) */
  imageBackBase64?: string;
}

/**
 * @nullable
 */
export type CardAnalysisResultHoloType = typeof CardAnalysisResultHoloType[keyof typeof CardAnalysisResultHoloType] | null;


export const CardAnalysisResultHoloType = {
  standard: 'standard',
  holo: 'holo',
  reverse_holo: 'reverse_holo',
} as const;

export interface CardAnalysisResult {
  cardId: number;
  /** @nullable */
  cardName?: string | null;
  /** @nullable */
  setName?: string | null;
  /** @nullable */
  cardNumber?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  holoType?: CardAnalysisResultHoloType;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  rarity?: string | null;
  /** Confidence score 0-1 */
  confidence?: number;
  /** @nullable */
  imageUrl?: string | null;
}

export type BatchAnalyzeInputImagesItem = {
  /** Base64-encoded front image */
  imageBase64: string;
  /** Base64-encoded back image (optional) */
  imageBackBase64?: string;
};

export interface BatchAnalyzeInput {
  images: BatchAnalyzeInputImagesItem[];
}

export interface BatchJob {
  jobId: string;
  total: number;
}

export type CardStatsHoloBreakdown = {
  standard: number;
  holo: number;
  reverse_holo: number;
};

export type CardStatsQualityBreakdownItem = {
  quality: string;
  count: number;
};

export interface CardStats {
  total: number;
  pending: number;
  reviewed: number;
  listed: number;
  holoBreakdown: CardStatsHoloBreakdown;
  qualityBreakdown: CardStatsQualityBreakdownItem[];
  recentlyAdded?: Card[];
}

export type PricingDataSoldListingsItem = {
  title: string;
  price: number;
  soldDate: string;
  url: string;
};

export interface PricingData {
  cardId: number;
  /** @nullable */
  averagePrice?: number | null;
  /** @nullable */
  lowestPrice?: number | null;
  /** @nullable */
  highestPrice?: number | null;
  /** @nullable */
  suggestedPrice?: number | null;
  soldListings: PricingDataSoldListingsItem[];
}

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  draft: 'draft',
  active: 'active',
  sold: 'sold',
  ended: 'ended',
} as const;

export interface Listing {
  id: number;
  cardId: number;
  card?: Card;
  /** @nullable */
  ebayListingId?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  status: ListingStatus;
  /** @nullable */
  ebayUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ListingInput {
  cardId: number;
  price?: number;
  title?: string;
  description?: string;
}

export interface BulkListingInput {
  cardIds: number[];
}

export interface ListingStats {
  total: number;
  draft: number;
  active: number;
  sold: number;
  ended: number;
  totalRevenue: number;
  recentListings?: Listing[];
}

export interface EbayStatus {
  connected: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface EbayAuthUrl {
  url: string;
}

export type ListCardsParams = {
status?: ListCardsStatus;
holoType?: ListCardsHoloType;
search?: string;
};

export type ListCardsStatus = typeof ListCardsStatus[keyof typeof ListCardsStatus];


export const ListCardsStatus = {
  pending: 'pending',
  reviewed: 'reviewed',
  listed: 'listed',
} as const;

export type ListCardsHoloType = typeof ListCardsHoloType[keyof typeof ListCardsHoloType];


export const ListCardsHoloType = {
  standard: 'standard',
  holo: 'holo',
  reverse_holo: 'reverse_holo',
} as const;

export type ImportEbayCsvBody = {
  csvContent: string;
};

export type GetImportProgress200 = {
  processed: number;
  total: number;
  done: boolean;
  imported: number;
  priced: number;
  errors: number;
};

export type ListListingsParams = {
status?: ListListingsStatus;
};

export type ListListingsStatus = typeof ListListingsStatus[keyof typeof ListListingsStatus];


export const ListListingsStatus = {
  draft: 'draft',
  active: 'active',
  sold: 'sold',
  ended: 'ended',
} as const;

export type EbayCallbackParams = {
code?: string;
state?: string;
};