# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CardLister API - trading card eBay listing tool
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List all cards
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, reviewed, listed]
        - name: holoType
          in: query
          schema:
            type: string
            enum: [standard, holo, reverse_holo]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: createCard
      tags: [cards]
      summary: Create card record manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
...
  /cards/analyze:
    post:
      operationId: analyzeCard
      tags: [cards]
      summary: Analyze a single card image with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardAnalyzeInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardAnalysisResult"
  /cards/batch-analyze:
    post:
      operationId: batchAnalyzeCards
      tags: [cards]
      summary: Batch analyze multiple card images with SSE progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchAnalyzeInput"
      responses:
        "200":
          description: Created batch job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchJob"
  /cards/batch-analyze/{jobId}/progress:
    get:
      operationId: getBatchProgress
      tags: [cards]
      summary: SSE stream for batch analysis progress
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
  /cards/{id}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get a card by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Not found
    patch:
      operationId: updateCard
      tags: [cards]
      summary: Update card details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
    delete:
      operationId: deleteCard
      tags: [cards]
      summary: Delete a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cards/{id}/pricing:
    get:
      operationId: getCardPricing
      tags: [cards]
      summary: Get eBay recently sold pricing for a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pricing data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingData"
...
components:
  schemas:
...
    Card:
      type: object
      required: [id, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        imageUrl:
          type: ["string", "null"]
        cardName:
          type: ["string", "null"]
        setName:
          type: ["string", "null"]
        cardNumber:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        quality:
          type: ["string", "null"]
          description: "NM, LP, MP, HP, or D"
        holoType:
          type: ["string", "null"]
          enum: ["standard", "holo", "reverse_holo", null]
        language:
          type: ["string", "null"]
        rarity:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, reviewed, listed]
        suggestedPrice:
          type: ["number", "null"]
        ebayListingId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CardInput:
      type: object
      properties:
        imageUrl:
          type: string
        cardName:
          type: string
        setName:
          type: string
        cardNumber:
          type: string
        year:
          type: string
        quality:
          type: string
        holoType:
          type: string
          enum: [standard, holo, reverse_holo]
        language:
          type: string
        rarity:
          type: string
        notes:
          type: string
        suggestedPrice:
          type: number
    CardUpdate:
      type: object
      properties:
        imageUrl:
          type: string
        cardName:
          type: string
        setName:
          type: string
        cardNumber:
          type: string
        year:
          type: string
        quality:
          type: string
        holoType:
          type: string
          enum: [standard, holo, reverse_holo]
        language:
          type: string
        rarity:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [pending, reviewed, listed]
        suggestedPrice:
          type: number
    CardAnalyzeInput:
      type: object
      required: [imageBase64]
      properties:
        imageBase64:
          type: string
          description: Base64-encoded front image data
        imageBackBase64:
          type: string
          description: Base64-encoded back image data (optional)
    CardAnalysisResult:
      type: object
      required: [cardId]
      properties:
        cardId:
          type: integer
        cardName:
          type: ["string", "null"]
        setName:
          type: ["string", "null"]
        cardNumber:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        quality:
          type: ["string", "null"]
        holoType:
          type: ["string", "null"]
          enum: ["standard", "holo", "reverse_holo", null]
        language:
          type: ["string", "null"]
        rarity:
          type: ["string", "null"]
        confidence:
          type: number
          description: Confidence score 0-1
        imageUrl:
          type: ["string", "null"]
    BatchAnalyzeInput:
      type: object
      required: [images]
      properties:
        images:
          type: array
          items:
            type: object
            required: [imageBase64]
            properties:
              imageBase64:
                type: string
                description: Base64-encoded front image
              imageBackBase64:
                type: string
                description: Base64-encoded back image (optional)
    BatchJob:
      type: object
      required: [jobId, total]
      properties:
        jobId:
          type: string
        total:
          type: integer
...
    PricingData:
      type: object
      required: [cardId, soldListings]
      properties:
        cardId:
          type: integer
        averagePrice:
          type: ["number", "null"]
        lowestPrice:
          type: ["number", "null"]
        highestPrice:
          type: ["number", "null"]
        suggestedPrice:
          type: ["number", "null"]
        soldListings:
          type: array
          items:
            type: object
            required: [title, price, soldDate, url]
            properties:
              title:
                type: string
              price:
                type: number
              soldDate:
                type: string
              url:
                type: string
...